+++ internal/storage/withdraw.go
@@ -4,8 +4,8 @@ package storage
 // TODO: add JSON and SQL specification
 //
 type Withdraw struct {
-	OrderNumber     string `json:"order" gorm:"unique"`
-	AccrualWithdraw uint   `json:"sum"`
+	OrderNumber     string  `json:"order" gorm:"unique"`
+	AccrualWithdraw float32 `json:"sum"`
 	UserID          uint
 	Status          string
 }
